Fixed data type issues

+++ AngularJSAuthentication.API/Dto/TemplateTaskDto.cs
@@ -17,8 +17,8 @@ namespace AngularJSAuthentication.API.Dto
             Tminus = templateTask.Tminus;
         }
 
-        public int TemplateId { get; set; }
-        public int TaskId { get; set; }
+        public long TemplateId { get; set; }
+        public long TaskId { get; set; }
         public string Stage { get; set; }
         public string Environment { get; set; }
         public string Domain { get; set; }
